Add files via upload
Task 1: Initialize an array of your favorite fruits. Add “Mango” to the end of the array 
using push().
Task 2: Remove the last fruit from the array using pop().
Task 3: Add “Strawberry” to the beginning of the fruits array using unshift().
Task 4: Remove the first fruit from the array using shift().
Task 5: Create a function that accepts an array of numbers and uses push() to add the number 7 to it.